backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
Pillow==10.1.0
django-filter==23.5
drf-yasg==1.21.7
python-dotenv==1.0.0
qrcode==7.4.2
reportlab==4.0.8
xlsxwriter==3.1.9
django-storages==1.14.2
django-import-export==3.3.3
django-celery-beat==2.5.0
celery==5.3.6
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0 